Add seed parameter to scan distortion functions in noise.py
The "_single_axis_distortion" and "_make_displacement_field" functions now accept a seed argument, to allow for deterministic randomness in their computation. The random number generator is now specifically initiated using the received seed. As well, the seed argument is passed through the scan noise function in "measurements.py" to the aforementioned functions.